Rename package to casper
