+++ src/casper/CasperBot.java
@@ -1,4 +1,4 @@
-package main.java;
+package casper;
 import exception.*;
 
 import java.time.LocalDate;
