+++ src/casper/Event.java
@@ -1,4 +1,4 @@
-package main.java;
+package casper;
 
 import java.time.LocalDate;
